Fix upstream sync (again) (#2763)
* sync from upstream with new files

* formatting
