+++ ci/tflite_files.txt
@@ -1,5 +1,7 @@
 tensorflow/compiler/mlir/lite/core/api/error_reporter.h
 tensorflow/compiler/mlir/lite/core/api/error_reporter.cc
+tensorflow/compiler/mlir/lite/core/c/builtin_op_data.h
+tensorflow/compiler/mlir/lite/core/c/tflite_types.h
 tensorflow/compiler/mlir/lite/schema/schema.fbs
 tensorflow/compiler/mlir/lite/schema/schema_utils.h
 tensorflow/compiler/mlir/lite/schema/schema_utils.cc
